Added a display field to the mouse state struct, to indicate the display the mouse is currently on, or NULL is the mouse is not actually in any of our windows at the moment. With help by Milan Mimica for the Windows code.
TODO: if a button is held down, we may not want to set the display field to
NULL until the button is released because we should register the button
down state only for the window the mouse was on when the button was
clicked. This requires a bit of bookkeeping.