Update font/ttf addons to use new UTF-8 string routines.
This lifts some arbitrary limits and arguably simplifies some of the code.

The `text_length' and `render' methods in ALLEGRO_FONT_VTABLE now take
ALLEGRO_USTR arguments.  `render' now returns the length of the rendered text
in pixels, like `text_length' which is useful for the implementation of
justified text.