al_get_stream_fragment may return NULL even if called in response to ALLEGRO_EVENT_STREAM_EMPTY_FRAGMENT.
Essentially the user is unreliable so Allegro does not know how many
events that it has emitted will be responded to.  Hence it must be
conservative, so there may be more fragment events than actual
fragments.  Make this clearer in the documentation.

Change the function signature of al_get_stream_fragment() to be more
straightforward, and hopefully encourage the user to check its return
value.