Speed up text log appending by making the reader/writers asynchronous.
The writer only needs to append text to a pending buffer, schedule the
reader to run (if it isn't already scheduled), then it can return
immediately.

When the reader eventually runs, it copies all of the pending text to
the underlying text widget and scrolls the widget.  With the GTK
implementation, the scrolling was the bottleneck.  Now many lines of
text may be appended before each scroll, so it is much faster.