x11: Use _NET_WM_STATE_FULLSCREEN hint for "true" fullscreen displays.
Previously we only used it for fullscreen-windows, but since it is
a cleaner way to tell modern window managers we want to fill the entire
screen, we can and should use it even if we change the video mode.
Tested with XFCE, KDE4, Metacity, IceWM, OpenBox.  Fluxbox mostly works
too (probably as well as it did before).  There are still some hacks, as
usual.

Re-enable mouse pointer grabbing in fullscreen modes.