Add al_is_system_installed and hide al_get_system_driver.
ALLEGRO_SYSTEM is not a documented type so the only way the user could
use al_get_system_driver() is testing if it returns NULL.
The function exists in 5.0 RC1 so we better not remove it, but we can
remove the documentation.

Add al_is_system_installed as a replacement, consistent with other
similar functions.