Don't send KEY_CHAR events for modifier and dead keys on X11, and don't send KEY_DOWN events for non-physical key events.
Set the keycode field in KEY_CHAR events to the code of the last key
pressed, as stated in the documentation, even if the char is due to a
compose sequence.

Get rid of historic behaviour where the unicode field is always
zero when Alt is held down.

cherry-pick https://alleg.svn.sourceforge.net/svnroot/alleg/allegro/branches/5.1@14236