Make al_get_opengl_fbo attempt to create an FBO if one doesn't exist.
This would seem to be more useful than the old behaviour, which was to
only return the FBO which had already been created.  If the user calls
the function, presumably he wants an FBO to exist.

This also improves compatibility with the old semantics of
al_get_opengl_fbo, in the presence of automatically freeing FBOs
that had been created by al_set_target_bitmap.